github: disallow libc 0.2.165
Summary:
X-link: https://github.com/facebook/buck2/pull/814

`libc` 0.2.165 was released recently and it breaks `sysinfo` crate for Mac.

`sysinfo` is fixed by https://github.com/GuillaumeGomez/sysinfo/pull/1393 which is released as v0.32.1.

But there are breaking changes compare to current sysinfo v0.30.11 we have internally and it needs to be migrated.

Reviewed By: ndmitchell, Will-MingLun-Li

Differential Revision: D66499693

fbshipit-source-id: 390c76f85ce431f39882345500593ae008664b80